增加下载图片，判断是否存在

+++ Sooxie/spiders/DownloadImageSpider.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+import os
+
+import Image
 import scrapy
 import re
 from Sooxie.domain.Sooxie import *
@@ -49,12 +52,34 @@ class SooXieSpider(scrapy.Spider):
         for image in mainimageslist:
             self.savemainimage(image)
 
+        self.change_image_620()
+
+    def change_image_620(self):
+        rootdir = "d:\sooxie"
+        excludes = ['csv', '620']
+        for root, dir, filenames in os.walk(rootdir):  # 三个参数：分别返回1.父目录 2.所有文件夹名字（不含路径） 3.所有文件名字
+            dir[:] = [d for d in dir if d not in excludes]
+            for filename in filenames:  # 输出文件信息
+                infile = os.path.join(root, filename)  # 输出文件路径信息
+                outfile = os.path.join("d:\\sooxie\\620", filename)  # 输出文件路径信息
+                if os.path.exists(outfile):
+                    continue
+                print filename
+                print u"文件夹信息:" + infile  # 输出文件路径信息
+                print u"输出文件夹信息:" + outfile  # 输出文件路径信息
+                im = Image.open(infile)
+                (x, y) = im.size  # read image size
+                x_s = 620  # define standard width
+                y_s = y * x_s / x  # calc height based on standard width
+                out = im.resize((x_s, y_s), Image.ANTIALIAS)  # resize image with high-quality
+                out.save(outfile)
+
     def saveimage(self, image):
         url = image.Url
         filename = "d:\\sooxie\\"
         filepath = ""
         # 将正则表达式编译成Pattern对象
-        pattern = re.compile(r'\d.*\.[A-Za-z]{3}')
+        pattern = re.compile(r'/(?:.(?!/))+$')
         # 使用Pattern匹配文本，获得匹配结果，无法匹配时将返回None
         match = re.search(pattern, url)
         if match is None:
@@ -63,10 +88,14 @@ class SooXieSpider(scrapy.Spider):
         else:
             print(url)
         newname = match.group(0)
+        newname = newname.replace("/", "")
+        newname = newname.replace("!bac", "")
         # print(newname)
         if newname:
             # 使用Match获得分组信息
             filepath = filename + newname
+            if os.path.exists(filepath):
+                return
             # print(url)
             # print(filepath)
             urllib.urlretrieve(url, filepath)
@@ -88,6 +117,8 @@ class SooXieSpider(scrapy.Spider):
         if newname:
             # 使用Match获得分组信息
             filepath = filename + newname
+            if os.path.exists(filepath):
+                return
             # print(url)
             # print(filepath)
             urllib.urlretrieve(url, filepath)
